+++ CHANGES.md
@@ -1,4 +1,14 @@
 ------------------------------------------------------------------------------
+2022-03-09
+
+The document details the changes between this version, local leetcode 0.9.3.1 and previous version, local leetcode 0.9.2.1.
+
+- add parser and printer for type `RandomTreeNode` (e.g. binary tree node with one random pointer to a node in the tree)
+
+- minor fixes on function documentation
+
+
+------------------------------------------------------------------------------
 2022-03-04
 
 The document details the changes between this version, local leetcode 0.9.2.1 and previous version, local leetcode 0.9.1.1.